+++ src/vscode/treeviews/Workspaces/FabricWorkspaceRoleAssignment.ts
@@ -1,10 +1,11 @@
 import * as vscode from 'vscode';
 
 import { FabricWorkspaceTreeItem } from './FabricWorkspaceTreeItem';
-import { iFabricApiItem, iFabricApiWorkspaceRoleAssignment } from '../../../fabric/_types';
+import { iFabricApiItem, iFabricApiWorkspaceRoleAssignment, iFabricApiWorkspaceRoleAssignmentRole } from '../../../fabric/_types';
 import { FabricApiService } from '../../../fabric/FabricApiService';
 import { FabricItem } from './FabricItem';
 import { ThisExtension } from '../../../ThisExtension';
+import { Helper } from '@utils/Helper';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
 export class FabricWorkspaceRoleAssignment extends FabricWorkspaceTreeItem {
@@ -14,8 +15,10 @@ export class FabricWorkspaceRoleAssignment extends FabricWorkspaceTreeItem {
 	) {
 		super(definition.id, definition.principal.displayName, "WorkspaceRoleAssignment", parent, definition, "", vscode.TreeItemCollapsibleState.None);
 
+		this.id = parent.id + "/" + definition.id;
 		this.label = definition.principal.displayName;
 		this.description = this._description;
+		this.contextValue = this._contextValue;
 
 		this.iconPath = {
 			light: this.getIconPath(),
@@ -38,6 +41,17 @@ export class FabricWorkspaceRoleAssignment extends FabricWorkspaceTreeItem {
 	}
 
 	/* Overwritten properties from FabricApiTreeItem */
+	get _contextValue(): string {
+		let orig: string = super._contextValue;
+
+		let actions: string[] = [
+			"DELETE",
+			"UPDATE"
+		];
+
+		return orig + actions.join(",") + ",";
+	}
+
 	get _description(): string {
 		if (this.itemDefinition?.principal) {
 			let desc = this.itemDefinition.role;
@@ -65,13 +79,47 @@ export class FabricWorkspaceRoleAssignment extends FabricWorkspaceTreeItem {
 	// properties of iFabricApiWorkspaceRoleAssignment
 	async delete(): Promise<void> {
 		try {
-			const result = await FabricApiService.awaitWithProgress("Deleting Role Assignment", FabricApiService.delete(this.apiPath, undefined));
+			const result = await FabricApiService.awaitWithProgress("Deleting Role Assignment", FabricApiService.delete(this.apiPath, undefined), 2000);
 
 			if (result.success) {
-				ThisExtension.Logger.logInfo(`Role Assignment '${this.itemId}' deleted from workspace '${this.workspace.itemName}'`);
+				ThisExtension.Logger.logInfo(`Role Assignment for '${this.itemDefinition.principal.displayName}' deleted from workspace '${this.workspace.itemName}'`);
+				ThisExtension.TreeViewWorkspaces.refresh(this.parent, false);
 			}
 			else {
-				ThisExtension.Logger.logError(`Could not delete Role Assignment '${this.itemId}' from workspace '${this.workspace.itemName}'`);
+				ThisExtension.Logger.logError(`Could not delete Role Assignment '${this.itemDefinition.principal.displayName}' from workspace '${this.workspace.itemName}'`);
+			}
+		}
+		catch (e) {
+			ThisExtension.Logger.logError(e.message);
+		}
+	}
+
+	async update(): Promise<void> {
+		const availableRoles = Helper.getQuickPicksFromEnum(iFabricApiWorkspaceRoleAssignmentRole, this.itemDefinition.role);
+		let role = await vscode.window.showQuickPick(availableRoles, {
+			"title": "Select new Role",
+			"placeHolder": this.itemDefinition.role.toString(),
+			"canPickMany": false
+		})
+
+		if(!role) {
+			ThisExtension.Logger.logWarning("No role selected. Aborting update of Role Assignment.");
+			return;
+
+		}
+		const body = {
+			"role": role.label
+		};
+
+		try {
+			const result = await FabricApiService.awaitWithProgress("Updating Role Assignment", FabricApiService.patch(this.apiPath, body), 2000);
+
+			if (result.success) {
+				ThisExtension.Logger.logInfo(`Role Assignment for '${this.itemDefinition.principal.displayName}' in workspace '${this.workspace.itemName}' updated to '${role}'.`);
+				ThisExtension.TreeViewWorkspaces.refresh(this.parent, false);
+			}
+			else {
+				ThisExtension.Logger.logError(`Could not update Role Assignment '${this.itemDefinition.principal.displayName}' in workspace '${this.workspace.itemName}'`);
 			}
 		}
 		catch (e) {