+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - added [Workspace Role Assignments](https://learn.microsoft.com/en-us/rest/api/fabric/core/workspaces/list-workspace-role-assignments?tabs=HTTP)
 - added `Run` actions for notebooks, data pipelines and Spark Jobs.
 - added viewer for Workspace Spark Settings
+- added button for Workspace Filter
+- added Drag & Drop provider for `WorkspaceRoleAssignments`
+- added `Update` and `Delete` actions for `WorkspaceRoleAssignments`
 
 **v0.9.8**:
 - added support for [Item Data Access Roles](https://learn.microsoft.com/en-us/rest/api/fabric/core/onelake-data-access-security/list-data-access-roles?tabs=HTTP)